Follow Coding Standard
The instinct is to code for functionality first, rather than readability.

Coding an incremental chatbot to allow for future changes requires standardisation, especially for different developers.

Let's follow the conventions laid out as per se-education.org to remain on the same page and allow for easy understanding and maintaining of code.

This helps not only with future or current differing developers, but also the main developer as code is often forgotten.