Add Level 6 - Delete
Added the delete function in Lemon.java to
allow deletion of tasks. Exception handling was
done for this delete function.

Updated the printAddedTask method to
printAddedOrDeletedTask method to print the
status of the list for both add and delete
functions.

Updated the checkEmptyList method to send
customized exception message for the delete
function in the case of an empty list.

Changed the fixed array used to store tasks to
a resizable ArrayList. This provides dynamic
sizing and improved flexibility. Updated
various methods to replace array indexing,
retrieving, adding and deleting of objects with
ArrayList methods.