Add A-MoreOOP
Created new classes Command.java, Messages.java,
Storage.java, TaskList.java, UI.java, Parser.java.

Extracted closely related functions as classes.

Command.java and *Command.java classes executes
the various commands from user.

Messages.java stores all the constant messages to
be displayed to the user.

Storage.java manages the loading and saving of
data to and from a text file.

TaskList.java manages the task list, it performs
functions like adding, deleting, marking, and
unmarking of tasks. Tasks in the task list and
size of the task list can also be retrieved here.
It can also check if the list is empty.

UI.java handles the receiving and displaying of
messages to and from the user.

Parser.java checks if the user input is valid and
of the right format before proceeding with the
execution of commands.